requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
nest-asyncio==1.5.8
aiohttp==3.9.1
pandas==2.1.4
plotly==5.18.0
crawl4ai
selenium
matplotlib
numpy
google-genai
